api endpoint for admin/recent objects and moved things around a little
