+++ application/models/models.go
@@ -42,8 +42,6 @@ const (
 
 	FieldClaimStatus        = "Status"
 	FieldItemCoverageStatus = "CoverageStatus"
-
-	QueryRecentStatusChanges = "created_at > ? AND field_name = ? AND action = ?"
 )
 
 type Authable interface {
